Splitt i to filer
